Refactor file structure in codemods package
