+++ packages/cli/src/presets/codemods/__tests__/codemods.test.tsx
@@ -22,7 +22,7 @@ const expectCodemodToHaveBeenRan = (name: string, runPath: string) => {
   expect(execAsync).toHaveBeenCalledWith(
     expect.stringMatching(
       new RegExp(
-        `.*--transform=.*node_modules\\/@compiled\\/codemods\\/(dist|src)\\/transforms\\/${name}\\/index.(tsx|js) ${regexPath}`
+        `.*--transform=.*node_modules\\/@compiled\\/codemods\\/(dist|src)\\/transforms\\/${name}\\/index.(ts|js) ${regexPath}`
       )
     )
   );
