--- packages/codemods/src/transforms/emotion-to-compiled/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './emotion-to-compiled';
