--- packages/codemods/src/transforms/styled-components-inner-ref-to-ref/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './styled-components-inner-ref-to-ref';
