--- packages/codemods/src/transforms/styled-components-to-compiled/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './styled-components-to-compiled';
